Move chad sort into its own lib, arg parse in scripts/, and move dirs operations to utility/utils_dirs.py
